iox-#27 Split test fixture of PortManager_test into separate hpp file
